@@ -159,7 +159,7 @@ export function prevDate(date, duration, hiddenDays) {
 export function getWeekNumber(date, firstDay) {
     // Copy date so don't modify original
     date = cloneDate(date);
-    if (firstDay == 0) {  // Western
+    if (firstDay === 0) {  // Western
         // Set to nearest Saturday: current date + 6 - current day number
         date.setUTCDate(date.getUTCDate() + 6 - date.getUTCDay());
     } else {  // ISO
