@@ -1,4 +1,6 @@
-import {addDay, addDuration, datesEqual, createDate, cloneDate, setMidnight, toLocalDate, toISOString, noTimePart, copyTime} from './date.js';
+import {
+    addDay, cloneDate, copyTime, createDate, datesEqual, noTimePart, setMidnight, toISOString, toLocalDate
+} from './date.js';
 import {createElement} from './dom.js';
 import {assign, isArray, isFunction} from './utils.js';
 import {toViewWithLocalDates} from './view.js';
@@ -55,38 +57,6 @@ export function createEventSources(input) {
     }));
 }
 
-export function createEventChunk(event, start, end) {
-    let chunk = {
-        start: event.start > start ? event.start : start,
-        end: event.end < end ? event.end : end,
-        event
-    };
-    chunk.zeroDuration = datesEqual(chunk.start, chunk.end);
-
-    return chunk;
-}
-
-export function sortEventChunks(chunks, eventOrder) {
-    if (isFunction(eventOrder)) {
-        chunks.sort((a, b) => eventOrder(
-            _toChunkWithLocalDates(a),
-            _toChunkWithLocalDates(b)
-        ));
-    } else {
-        // Sort by start date (all-day events always on top)
-        chunks.sort((a, b) => a.start - b.start || b.event.allDay - a.event.allDay);
-    }
-}
-
-function _toChunkWithLocalDates(chunk) {
-    return {
-        ...chunk,
-        start: toLocalDate(chunk.start),
-        end: toLocalDate(chunk.end),
-        event: toEventWithLocalDates(chunk.event)
-    };
-}
-
 export function createEventContent(chunk, displayEventEnd, eventContent, theme, _intlEventTime, _view) {
     let timeText = _intlEventTime.formatRange(
         chunk.start,
@@ -113,7 +83,7 @@ export function createEventContent(chunk, displayEventEnd, eventContent, theme, 
                 domNodes = [];
                 break;
             case 'pointer':
-                domNodes = [createTimeElement(timeText, chunk, theme)];
+                domNodes = chunk.event.allDay ? [] : [createTimeElement(timeText, chunk, theme)];
                 break;
             default:
                 domNodes = [
@@ -169,118 +139,24 @@ function _cloneEvent(event, dateFn) {
     return event;
 }
 
-/**
- * Prepare event chunks for month view and all-day slot in week view
- */
-export function prepareEventChunks(chunks, hiddenDays, eventOrder) {
-    let longChunks = {};
-
-    if (chunks.length) {
-        sortEventChunks(chunks, eventOrder);
-
-        let prevChunk;
-        for (let chunk of chunks) {
-            let dates = [];
-            let date = setMidnight(cloneDate(chunk.start));
-            while (chunk.end > date) {
-                if (!hiddenDays.includes(date.getUTCDay())) {
-                    dates.push(cloneDate(date));
-                    if (dates.length > 1) {
-                        let key = date.getTime();
-                        if (longChunks[key]) {
-                            longChunks[key].chunks.push(chunk);
-                        } else {
-                            longChunks[key] = {
-                                sorted: false,
-                                chunks: [chunk]
-                            };
-                        }
-                    }
-                }
-                addDay(date);
-            }
-            if (dates.length) {
-                chunk.date = dates[0];
-                chunk.days = dates.length;
-                chunk.dates = dates;
-                // Adjust the start and end dates of the chunk if hidden days affected them
-                if (chunk.start < dates[0]) {
-                    chunk.start = dates[0];
-                }
-                let maxEnd = addDay(cloneDate(dates.at(-1)));
-                if (chunk.end > maxEnd) {
-                    chunk.end = maxEnd;
-                }
-            } else {
-                chunk.date = setMidnight(cloneDate(chunk.start));
-                chunk.days = 1;
-                chunk.dates = [chunk.date];
-            }
-
-            if (prevChunk && datesEqual(prevChunk.date, chunk.date)) {
-                chunk.prev = prevChunk;
-            }
-            prevChunk = chunk;
-        }
-    }
-
-    return longChunks;
-}
-
-export function repositionEvent(chunk, longChunks, height) {
-    chunk.top = 0;
-    if (chunk.prev) {
-        chunk.top = chunk.prev.bottom + 1;
-    }
-    chunk.bottom = chunk.top + height;
-    let margin = 1;
-    let key = chunk.date.getTime();
-    if (longChunks[key]) {
-        if (!longChunks[key].sorted) {
-            longChunks[key].chunks.sort((a, b) => a.top - b.top);
-            longChunks[key].sorted = true;
-        }
-        for (let longChunk of longChunks[key].chunks) {
-            if (chunk.top < longChunk.bottom && chunk.bottom > longChunk.top) {
-                let offset = longChunk.bottom - chunk.top + 1;
-                margin += offset;
-                chunk.top += offset;
-                chunk.bottom += offset;
-            }
-        }
-    }
-
-    return margin;
-}
-
 export function runReposition(refs, data) {
     refs.length = data.length;
-    let result = [];
     for (let ref of refs) {
-        result.push(ref?.reposition?.());
+        ref?.reposition();
     }
-    return result;
 }
 
 /**
- * Check whether the event intersects with the given date range and resources
+ * Check whether the event intersects with the given date range and resource
  * @param event
  * @param start
  * @param end
- * @param resources
+ * @param resource
  * @return boolean
  */
-export function eventIntersects(event, start, end, resources) {
-    if (event.start < end && event.end > start) {
-        if (resources) {
-            if (!isArray(resources)) {
-                resources = [resources];
-            }
-            return resources.some(resource => event.resourceIds.includes(resource.id));
-        }
-        return true;
-    }
-    return false;
+export function eventIntersects(event, start, end, resource = undefined) {
+    return (!resource || event.resourceIds.includes(resource.id)) && event.start < end && event.end > start;
+
 }
 
 export function helperEvent(display) {
