@@ -1,34 +1,40 @@
 <script>
     import {getContext} from 'svelte';
-    import {ancestor, bgEvent, height, max, rect, repositionEvent} from '#lib';
+    import {bgEvent, height, isEmpty, max, repositionEvent} from '#lib';
     import {InteractableEvent} from '#components';
 
-    let {chunk, longChunks = {}, inPopup = false, dates = []} = $props();
+    let {chunk, gridEl, inPopup = false} = $props();
 
-    let {dayMaxEvents, _hiddenEvents, _popupDate} = getContext('state');
+    let {_colsCount, _hiddenChunks, _popupDay, dayMaxEvents} = getContext('state');
 
     let el = $state();
-    let margin = $state(1);
+    let margin = $state(0);
     let hidden = $state(false);
 
     let event = $derived(chunk.event);
     let display = $derived(chunk.event.display);
+    let dayEl = $derived(gridEl?.children.item((chunk.gridRow - 1) * $_colsCount + chunk.gridColumn - 1));
+
+    $effect(() => {
+        if (dayEl) {
+            margin = height(dayEl.firstElementChild);
+        }
+    });
 
     // Style
     let styles = $derived(style => {
-        if (bgEvent(display)) {
-            style['width'] = `calc(${chunk.days * 100}% + ${(chunk.days - 1)}px)`;
-        } else {
-            let marginTop = margin;
+        style['grid-column'] = `${chunk.gridColumn} / span ${chunk.dates.length}`;
+        style['grid-row'] = chunk.gridRow;
+        if (!bgEvent(display)) {
+            let marginTop = inPopup ? 1 : margin;
             if (event._margin) {
                 // Force margin for helper events
-                let [_margin, _dates] = event._margin;
-                if (chunk.date >= _dates[0] && chunk.date <= _dates.at(-1)) {
+                let [_margin, _gridRow] = event._margin;
+                if (_margin > marginTop && chunk.gridRow === _gridRow) {
                     marginTop = _margin;
                 }
             }
-            style['width'] = `calc(${chunk.days * 100}% + ${(chunk.days - 1) * 7}px)`;
-            style['margin-top'] = `${marginTop}px`;
+            style['margin-block-start'] = `${marginTop}px`;
         }
         if (hidden) {
             style['visibility'] = 'hidden';
@@ -37,42 +43,37 @@
     });
 
     export function reposition() {
-        margin = repositionEvent(chunk, longChunks, height(el));
-        if ($dayMaxEvents === true) {
-            hide();
-        } else {
-            hidden = false;
-        }
+        margin = repositionEvent(chunk, height(el), height(dayEl.firstElementChild));
     }
 
-    function hide() {
-        let dayEl = ancestor(el, 2);
-        let h = height(dayEl) - height(dayEl.firstElementChild) - footHeight(dayEl);
-        hidden = chunk.bottom > h;
-        let update = false;
-        // Hide or show the event throughout all days
-        for (let date of chunk.dates) {
-            let hiddenEvents = $_hiddenEvents[date.getTime()];
-            if (hiddenEvents) {
-                let size = hiddenEvents.size;
-                if (hidden) {
-                    hiddenEvents.add(chunk.event);
-                } else {
-                    hiddenEvents.delete(chunk.event);
-                }
-                if (size !== hiddenEvents.size) {
-                    update = true;
+    export function hide() {
+        if ($dayMaxEvents === true) {
+            let h = height(dayEl) - footHeight(dayEl);
+            hidden = chunk.bottom > h;
+            if (hidden) {
+                // Hide the event throughout all days
+                for (let date of chunk.dates) {
+                    let key = date.getTime();
+                    if ($_hiddenChunks[key]) {
+                        if (!$_hiddenChunks[key].includes(chunk)) {
+                            $_hiddenChunks[key] = [...$_hiddenChunks[key], chunk];
+                        }
+                    } else {
+                        $_hiddenChunks[key] = [chunk];
+                    }
                 }
             }
-        }
-        if (update) {
-            $_hiddenEvents = $_hiddenEvents;
+        } else {
+            hidden = false;
+            if (!isEmpty($_hiddenChunks)) {
+                $_hiddenChunks = {};
+            }
         }
     }
 
     function footHeight(dayEl) {
         let h = 0;
-        for (let i = 0; i < chunk.days; ++i) {
+        for (let i = 0; i < chunk.dates.length; ++ i) {
             h = max(h, height(dayEl.lastElementChild));
             dayEl = dayEl.nextElementSibling;
             if (!dayEl) {
@@ -88,6 +89,6 @@
     {chunk}
     {styles}
     axis="x"
-    forceDate={() => inPopup ? $_popupDate : undefined}
-    forceMargin={() => [rect(el).top - rect(ancestor(el, 1)).top, dates]}
+    forceDate={inPopup && $_popupDay.dayStart}
+    forceMargin={[margin, chunk.gridRow]}
 />
