@@ -4,46 +4,37 @@
     import {repositionEvent} from './lib.js';
     import {InteractableEvent} from '#components';
 
-    let {chunk, dayChunks = [], longChunks = {}, resource = undefined} = $props();
+    let {chunk} = $props();
 
-    let {slotDuration, slotWidth} = getContext('state');
+    let {slotDuration, slotWidth, _monthView} = getContext('state');
 
     let el = $state();
-    let margin = $state(helperEvent(chunk.event.display) ? 1 : 0);
+    let margin = $state(1);
     let event = $derived(chunk.event);
     // Style
-    let width = $derived('slots' in chunk ? chunk.slots * $slotWidth : chunk.days * 100);
     let styles = $derived(style => {
-        if ('slots' in chunk) {
-            let left = chunk.offset * $slotWidth;
-            style['left'] = `${left}px`;
-            style['width'] = `${width}px`;
-        } else {
-            // Month view
-            style['width'] = `${width}%`;
+        style['grid-column'] = `${chunk.gridColumn} / span ${chunk.dates.length}`;
+        style['grid-row'] = chunk.gridRow;
+        if (!$_monthView) {
+            let left = chunk.left / toSeconds($slotDuration) * $slotWidth;
+            style['inset-inline-start'] = `${left}px`;
+            style['inline-size'] = `${chunk.width / toSeconds($slotDuration) * $slotWidth + chunk.dates.length - 1}px`;
         }
         let marginTop = margin;
         if (event._margin) {
             // Force margin for helper events
-            let [_margin, _resource] = event._margin;
-            if (resource === _resource) {
+            let [_margin, _gridRow] = event._margin;
+            if (chunk.gridRow === _gridRow) {
                 marginTop = _margin;
             }
         }
-        style['margin-top'] = `${marginTop}px`;
+        style['margin-block-start'] = `${marginTop}px`;
         return style;
     });
 
     export function reposition() {
-        if (!el) {
-            return 0;
-        }
-        let h = height(el);
-        margin = repositionEvent(chunk, dayChunks, longChunks, h, !toSeconds($slotDuration));
-        return margin + h;
+        margin = repositionEvent(chunk, height(el), $_monthView);
     }
 </script>
 
-{#if width > 0}
-    <InteractableEvent bind:el {chunk} {styles} axis="x" forceMargin={() => [margin, resource]} />
-{/if}
+<InteractableEvent bind:el {chunk} {styles} axis="x" forceMargin={[margin, chunk.gridRow]} />
