@@ -1,12 +1,12 @@
 <script>
     import {getContext} from 'svelte';
-    import {datesEqual, setContent, toISOString, toSeconds, observeResize} from '#lib';
+    import {datesEqual, contentFrom, toISOString, toSeconds, resizeObserver} from '#lib';
 
     let {_headerEl, _headerHeight, _intlDayHeader, _intlDayHeaderAL, _dayTimes, _today, _viewDates,
         slotDuration, theme} = getContext('state');
 </script>
 
-<div class="{$theme.header}" bind:this={$_headerEl} use:observeResize={() => $_headerHeight = $_headerEl.clientHeight}>
+<div class="{$theme.header}" bind:this={$_headerEl} {@attach resizeObserver(() => $_headerHeight = $_headerEl.clientHeight)}>
     <div class="{$theme.days}" role="row">
         {#each $_viewDates as date}
             <div class="{$theme.day} {$theme.weekdays?.[date.getUTCDay()]}{datesEqual(date, $_today) ? ' ' + $theme.today : ''}">
@@ -15,7 +15,7 @@
                         <time
                             datetime="{toISOString(date, 10)}"
                             aria-label="{$_intlDayHeaderAL.format(date)}"
-                            use:setContent={$_intlDayHeader.format(date)}
+                            {@attach contentFrom($_intlDayHeader.format(date))}
                         ></time>
                     </div>
                     <div class="{$theme.times}">
@@ -23,7 +23,7 @@
                             <div class="{$theme.time}{time[2] ? '' : ' ' + $theme.minor}" role="columnheader">
                                 <time
                                     datetime="{time[0]}"
-                                    use:setContent={time[1]}
+                                    {@attach contentFrom(time[1])}
                                 ></time>
                             </div>
                         {/each}
@@ -33,7 +33,7 @@
                         <time
                             datetime="{toISOString(date, 10)}"
                             aria-label="{$_intlDayHeaderAL.format(date)}"
-                            use:setContent={$_intlDayHeader.format(date)}
+                            {@attach contentFrom($_intlDayHeader.format(date))}
                         ></time>
                     </div>
                 {/if}
